synthesizeTopology: move to TopologyUtil, rename
