+++ projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/OspfNeighborTest.java
@@ -9,7 +9,7 @@ import static org.hamcrest.Matchers.not;
 
 import com.google.common.collect.ImmutableMap;
 import java.util.Map;
-import org.batfish.common.util.CommonUtil;
+import org.batfish.common.topology.TopologyUtil;
 import org.batfish.datamodel.ospf.OspfArea;
 import org.batfish.datamodel.ospf.OspfProcess;
 import org.batfish.datamodel.ospf.OspfTopologyUtils;
@@ -68,7 +68,7 @@ public class OspfNeighborTest {
 
     Map<String, Configuration> configurations =
         ImmutableMap.of(c1.getHostname(), c1, c2.getHostname(), c2);
-    Topology topology = CommonUtil.synthesizeTopology(configurations);
+    Topology topology = TopologyUtil.synthesizeL3Topology(configurations);
     OspfTopologyUtils.initRemoteOspfNeighbors(configurations, topology);
 
     assertThat(op1, hasOspfNeighbors(hasKey(new IpLink(ip1, Ip.ZERO))));
