+++ projects/batfish/src/test/java/org/batfish/representation/aws/RdsInstanceTest.java
@@ -27,6 +27,7 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import org.batfish.common.topology.TopologyUtil;
 import org.batfish.common.util.CommonUtil;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.Edge;
@@ -91,7 +92,7 @@ public class RdsInstanceTest {
   @Test
   public void testRdsSubnetEdge() throws IOException {
     Map<String, Configuration> configurations = loadAwsConfigurations();
-    Topology topology = CommonUtil.synthesizeTopology(configurations);
+    Topology topology = TopologyUtil.synthesizeL3Topology(configurations);
 
     // check that RDS instance is a neighbor of both  subnets in which its interfaces are
     assertThat(
